user_service/requirements.txt
fastapi
uvicorn
sqlalchemy 
python-jose 
passlib[argon2]
python-dotenv
pydantic[email]
alembic
sqlalchemy[asyncio] 
asyncpg